bug fix in delete method from phd candidacy referee letter

+++ src/net/sourceforge/fenixedu/domain/phd/candidacy/PhdCandidacyRefereeLetter.java
@@ -78,7 +78,9 @@ public class PhdCandidacyRefereeLetter extends PhdCandidacyRefereeLetter_Base {
     }
 
     public void delete() {
-	getFile().delete();
+	if (hasFile()) {
+	    getFile().delete();
+	}
 	removeRefereeCountry();
 	removeCandidacyReferee();
 	removePhdProgramCandidacyProcess();
